Move game_items::Item unit tests to ItemTests.cpp
